Fixes for build warnings on Windows. Fix PKCS7 to use const for byte array declaration. Cleanup of the pkcs7 MAX_PKCS7_DIGEST_SZ. Fix for unsigned / signed comparison warning for pkcs7_load_certs_keys in test.c. Fix for cast warning from word16 to byte in asn.c. Fix for build error with io.h refactor for InTime RTOS.

+++ wolfcrypt/src/asn.c
@@ -1049,19 +1049,19 @@ static word32 SetBitString16Bit(word16 val, byte* output)
 
     if ((val >> 8) != 0) {
         len = 2;
-        lastByte = val >> 8;
+        lastByte = (byte)(val >> 8);
     }
     else {
         len = 1;
-        lastByte = val;
+        lastByte = (byte)val;
     }
 
     while (((lastByte >> unusedBits) & 0x01) == 0x00)
         unusedBits++;
 
     idx = SetBitString(len, unusedBits, output);
-    output[idx++] = val;
-    output[idx++] = val >> 8;
+    output[idx++] = (byte)val;
+    output[idx++] = (byte)(val >> 8);
 
     return idx;
 }
